feat: implement image locality as score plugin
